Refactor: GoogleMapModal.tsx의 지도 동작부분과 <div>요소를 하나의 컴포넌트로 뺌
