+++ okivery/src/components/address/GoogleMapModal.tsx
@@ -1,20 +1,12 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useState } from "react";
 import "./GoogleMapModal.css";
-import { Loader } from "@googlemaps/js-api-loader";
-import { PacmanLoader } from "react-spinners";
-import isWithinOneKm from "./CalculateDistance";
-import centerLocation from "../../constants/location";
+import { PacmanLoader as LoadingIcon } from "react-spinners";
+import SelectLocationMap from "./SelectLocationMap";
 
 interface MapModalProps {
   onClose: () => void;
   onSelectAddress: (selectedAddress: string) => void;
   setIsAvailableService: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
-declare global {
-  interface Window {
-    google: typeof google;
-  }
 }
 
 const GoogleMapModal: React.FC<MapModalProps> = ({
@@ -23,110 +15,9 @@ const GoogleMapModal: React.FC<MapModalProps> = ({
   setIsAvailableService,
 }) => {
   const [isLoading, setIsLoading] = useState<boolean>(true);
-  const mapRef = useRef<HTMLDivElement>(null);
+  // 사용자 입력 주소가 배달가능한 지역인지
   const [isAvailable, setIsAvailable] = useState<boolean>(false);
   const [userAddressData, setUserAddressData] = useState<string>("");
-
-  useEffect(() => {
-    const loader = new Loader({
-      apiKey: import.meta.env.VITE_APP_GOOGLE_MAP_API_KEY,
-      language: "en",
-    });
-
-    initLibrary();
-
-    async function initLibrary() {
-      const infoWindow = new google.maps.InfoWindow();
-      const geocoder = new google.maps.Geocoder();
-      // 고급 마커
-      const { AdvancedMarkerElement } = await loader.importLibrary("marker");
-
-      navigator.geolocation.getCurrentPosition(
-        (position) => {
-          const userLatLng = {
-            lat: position.coords.latitude,
-            lng: position.coords.longitude,
-          };
-          // 지도 그리기
-          const map = new google.maps.Map(mapRef.current!, {
-            center: userLatLng,
-            zoom: 15,
-            mapId: "DEMO_MAP_ID",
-          });
-          // 마커 그리기
-          const marker = new AdvancedMarkerElement({
-            position: userLatLng,
-            map,
-            gmpDraggable: true,
-          });
-
-          // 사용자 현재 위치 주소정보 주소창으로 표시
-          geocoder.geocode(
-            { location: userLatLng, language: "en" },
-            (results, status) => {
-              if (status === "OK") {
-                if (results && results.length > 0) {
-                  const englishAddress = results[0].formatted_address;
-                  setUserAddressData(englishAddress);
-                  infoWindow.setContent(englishAddress);
-                  infoWindow.open(map, marker);
-                  setIsAvailable(
-                    isWithinOneKm(
-                      centerLocation.lat,
-                      centerLocation.lng,
-                      userLatLng.lat,
-                      userLatLng.lng
-                    )
-                  );
-                } else {
-                  console.log("No results found");
-                }
-              } else {
-                console.error("Geocoder failed due to: " + status);
-              }
-            }
-          );
-          setIsLoading(false);
-
-          // 마커를 드래그하다 놓은 위치의 위도 경도를 변환해 영문 주소로 바꾸고 주소정보창을 띄움
-          marker.addListener("dragend", () => {
-            const newPosition = marker.position;
-            if (newPosition) {
-              setIsAvailable(
-                isWithinOneKm(
-                  centerLocation.lat,
-                  centerLocation.lng,
-                  Number(newPosition.lat),
-                  Number(newPosition.lng)
-                )
-              );
-              geocoder.geocode(
-                { location: newPosition, language: "en" },
-                (results, status) => {
-                  if (status === "OK") {
-                    if (results && results.length > 0) {
-                      const englishAddress = results[0].formatted_address;
-                      setUserAddressData(englishAddress);
-                      infoWindow.setContent(englishAddress);
-                      infoWindow.open(map, marker);
-                    } else {
-                      console.log("No results found");
-                    }
-                  } else {
-                    console.error("Geocoder failed due to: " + status);
-                  }
-                }
-              );
-            }
-          });
-        },
-        (error) => {
-          console.error("Error getting user's location: ", error);
-          setIsLoading(false);
-        }
-      );
-    }
-  }, []);
 
   const handleSelectAddress = () => {
     setIsAvailableService(isAvailable);
@@ -146,10 +37,14 @@ const GoogleMapModal: React.FC<MapModalProps> = ({
           <div className="googleMap">
             {isLoading && (
               <div className="loadingSpinner">
-                <PacmanLoader size="50px" color="#ff6347" />
+                <LoadingIcon size="50px" color="#ff6347" />
               </div>
             )}
-            <div id="map" ref={mapRef} />
+            <SelectLocationMap
+              setIsLoading={setIsLoading}
+              setIsAvailable={setIsAvailable}
+              setUserAddressData={setUserAddressData}
+            />
           </div>
           <div className="saveButtonSection">
             <button
